Bug in Dashboard button fixed
